Refactor localization codes to an enum for type safety.
