--- solo-leveling-player-service/src/main/kotlin/com/sleepkqq/sololeveling/player/lozalization/LocalizationCodes.kt
@@ -1,7 +1,0 @@
-package com.sleepkqq.sololeveling.player.lozalization
-
-object LocalizationCodes {
-	const val TASKS_GENERATION_SUCCESS = "tasks.generation.success"
-	const val TABLES_PLAYER_BALANCE_TRANSACTIONS = "tables.player.balance.transactions"
-	const val TABLES_PLAYER_TASKS = "tables.player.tasks"
-}
