+++ solo-leveling-player-service/src/main/kotlin/com/sleepkqq/sololeveling/player/service/i18n/I18nService.kt
@@ -1,5 +1,6 @@
 package com.sleepkqq.sololeveling.player.service.i18n
 
+import com.sleepkqq.sololeveling.player.lozalization.LocalizationCode
 import org.springframework.context.MessageSource
 import org.springframework.context.i18n.LocaleContextHolder
 import org.springframework.stereotype.Service
@@ -9,8 +10,8 @@ class I18nService(
 	private val messageSource: MessageSource
 ) {
 
-	fun getMessage(key: String, vararg args: Any?): String = messageSource.getMessage(
-		key,
+	fun getMessage(code: LocalizationCode, vararg args: Any?): String = messageSource.getMessage(
+		code.code,
 		if (args.isEmpty()) null else args,
 		LocaleContextHolder.getLocale()
 	)
